Use the <nomodeline> argument in all 'doautocmd' invocations
This stops any modelines from being processed after the autocommands. As
we're only responding to events that shouldn't modify the buffer this is
both recommended per the Vim documentation (autocmd.txt : Section 9) and
avoids an edge case(s?) where re-applying modelines causes undesirable
behaviour. It may also improve performance (though no idea how much!).

With respect to the edge case: if the settings in a modeline are being
overridden by an autocommand, they won't work without the <nomodeline>
argument. This is because the modeline settings will always be applied
*after* the autocommands have been run.