Retry s:delayed_checktime on 'E523: Not allowed'
On neovim, when using vim-tmux-focus-events with vim-airline, the error
message:

  Error detected while processing function <SNR>52_delayed_checktime:
  line    8:
  E523: Not allowed here:   silent checktime

Appears after the following steps:
1. Open neovim in tmux,
2. Type a command on the command line (e.g. `:asdfasdfasdf`) without
pressing enter,
3. Focus a different window (e.g. click on a Chrome window),
4. Click on the neovim/tmux terminal window again,
5. Press <ESC> to quit the command line.

This seems to occur despite the plugin already checking whether the user
is in the command line (e.g. `mode() !=# 'c'` returns true in the
affected function, even when the error is thrown).

Wrap the `:silent checktime` call in s:delayed_checktime in a try/catch;
if an E523 is thrown, fail silently, and retry later.